Calendar Agent/requirements.txt
icalendar
python-dateutil
pytz
pandas
requests
PyYAML
